Fix previewnet bugs (#1787)
* Initial commit


* Disable CI for now


* Reenable CI


* Fix remaining known issues


* Fix typo


* Fix unit test


* Stabilize CI; remove refactored TokenStore.wipe() method


* Address reviewer comments; remove auxiliary id set from BackingNfts to reduce memory footprint


* Replace uses of ACCOUNT_AMOUNT_TRANSFERS_ONLY_ALLOWED_FOR_FUNGIBLE_COMMON


* Fix MerkleUniqueTokenId.deserialize()


* Minor changes.


* Stabilize CI

+++ hedera-node/src/test/java/com/hedera/services/state/merkle/MerkleUniqueTokenIdTest.java
@@ -45,8 +45,8 @@ class MerkleUniqueTokenIdTest {
 
 	private EntityId tokenId = MISSING_ENTITY_ID;
 	private EntityId otherTokenId = MISSING_ENTITY_ID;
-	private int serialNumber;
-	private int otherSerialNumber;
+	private long serialNumber;
+	private long otherSerialNumber;
 
 	@BeforeEach
 	void setup() {
@@ -130,7 +130,7 @@ class MerkleUniqueTokenIdTest {
 		SerializableDataInputStream in = mock(SerializableDataInputStream.class);
 
 		given(in.readSerializable()).willReturn(tokenId);
-		given(in.readInt()).willReturn(serialNumber);
+		given(in.readLong()).willReturn(serialNumber);
 
 		// and:
 		var read = new MerkleUniqueTokenId();
